refactor(opencode): share model identity rules
