+++ src/features/runtime-provider-management/renderer/hooks/openCodePassiveCatalogNormalization.ts
@@ -1,15 +1,13 @@
-import { parseOpenCodeQualifiedModelRef } from '@shared/utils/opencodeModelRef';
+import {
+  parseStrictQualifiedModelRef,
+  qualifyModelId,
+} from '../../core/domain/openCodeModelIdentity';
 
 import type {
   CliProviderModelAvailability,
   CliProviderModelCatalogItem,
   CliProviderStatus,
 } from '@shared/types';
-
-function parseStrictQualifiedModelRef(modelId: string | null | undefined) {
-  const parsed = parseOpenCodeQualifiedModelRef(modelId);
-  return parsed && parsed.raw.split('/').every((segment) => segment.length > 0) ? parsed : null;
-}
 
 export function normalizeProviderIdentity(
   providerId: string | null | undefined
@@ -18,28 +16,6 @@ export function normalizeProviderIdentity(
   return normalized && parseStrictQualifiedModelRef(`${normalized}/model`)?.sourceId === normalized
     ? normalized
     : null;
-}
-
-function qualifyModelId(providerId: string, modelId: string): string {
-  const normalizedProviderId = providerId.trim().toLowerCase();
-  const normalizedModelId = modelId.trim();
-  if (
-    !normalizedProviderId ||
-    !normalizedModelId ||
-    normalizedModelId !== modelId ||
-    /\s/.test(normalizedModelId)
-  ) {
-    return '';
-  }
-  if (normalizedModelId.includes('/')) {
-    const parsed = parseStrictQualifiedModelRef(normalizedModelId);
-    if (!parsed) {
-      return '';
-    }
-    return parsed.sourceId === normalizedProviderId ? parsed.raw : '';
-  }
-  const parsed = parseStrictQualifiedModelRef(`${normalizedProviderId}/${normalizedModelId}`);
-  return parsed?.sourceId === normalizedProviderId ? parsed.raw : '';
 }
 
 export function normalizePassiveCatalogModel(
